refactor(assets): extract useSendForm from SendForm.vue (#185)
Move the SendForm script verbatim into a co-located useSendForm composable
so the SFC is a thin projection of already-computed values, per the house
Vue style. Behaviour-preserving: data/derivation/validation/submit logic is
moved unchanged; the two inline template handlers become named composable
functions (onSelectCurrency, onRecipientInput).

Send and Receive get their own composables rather than one shared module:
their mutable non-reactive locals (client, route, chains) and the source/
sink asymmetry in getRoute are load-bearing and do not share without a
rewrite, which this behaviour-preserving pass avoids.
