+++ src/modules/assets/components/SendForm.vue
@@ -31,11 +31,7 @@
       :calculated-balance="calculatedBalance"
       :pickerPlacehodler="$t('message.loading')"
       :value="amount"
-      @on-selected-currency="
-        (option) => {
-          selectedCurrency = assets.findIndex((item) => item == option);
-        }
-      "
+      @on-selected-currency="onSelectCurrency"
       @input="handleAmountChange"
       :is-loading-picker="disablePicker"
       :disabled-input-field="isDisabled"
@@ -79,11 +75,7 @@
         :inputClass="network.native ? '' : 'border-none p-0'"
         :disabled="!network.native"
         :value="walletRef"
-        @input="
-          (e) => {
-            receiverAddress = (e.target as HTMLInputElement).value;
-          }
-        "
+        @input="onRecipientInput"
       />
     </div>
 
@@ -105,690 +97,26 @@
 </template>
 
 <script lang="ts" setup>
-import type { AssetBalance } from "@/common/stores/wallet/types";
-import type { Coin } from "@keplr-wallet/types";
-
-import { SwapStatus } from "../enums";
 import { AdvancedFormControl, Button, AssetItem, Input, type AssetItemProps } from "web-components";
-import { NETWORK_DATA } from "@/networks/config";
+import { h } from "vue";
 import { NATIVE_NETWORK } from "../../../config/global/network";
-import { IGNORED_NETWORKS } from "../../../config/global";
-
-import type { Wallet } from "@/networks";
-import { type BaseWallet } from "@/networks";
-import { CONFIRM_STEP, type ExternalCurrency, type Network, type SkipRouteConfigType } from "@/common/types";
-import { useWalletStore } from "@/common/stores/wallet";
-import { useBalancesStore } from "@/common/stores/balances";
-import { useHistoryStore } from "@/common/stores/history";
-import { computed, onUnmounted, ref, watch, h, inject } from "vue";
-import { useI18n } from "vue-i18n";
-import {
-  classifyError,
-  externalWallet,
-  Logger,
-  transferCurrency,
-  validateAddress,
-  walletOperation,
-  WalletAccess
-} from "@/common/utils";
-import { getSkipRouteConfig } from "@/common/utils/ConfigService";
-import { formatDecAsUsd, formatUsd, formatTokenBalance } from "@/common/utils/NumberFormatUtils";
-import { getPriceForCurrency, tryGetCurrencyByDenom } from "@/common/utils/CurrencyLookup";
-import { coin } from "@cosmjs/stargate";
-import { Decimal } from "@cosmjs/math";
-import { SkipRouter, type SkipTxResult } from "@/common/utils/SkipRoute";
-import type { NetworkInfo } from "@/common/api/types/config";
-import { Dec } from "@keplr-wallet/unit";
-import { ErrorCodes } from "@/config/global";
-import { useConfigStore } from "@/common/stores/config";
-import { HISTORY_ACTIONS } from "@/modules/history/types";
-import type { Chain, RouteResponse } from "@/common/types/skipRoute";
-
-const i18n = useI18n();
-
-const assets = computed(() => {
-  const data = [];
-  for (const asset of (networkCurrenciesRef.value as ExternalCurrency[] | AssetBalance[]) ?? []) {
-    const denom = (asset as ExternalCurrency).ibcData ?? (asset as AssetBalance).from;
-    const currency = tryGetCurrencyByDenom(denom);
-    if (!currency) continue; // Skip deprecated/removed denoms
-
-    const value = new Dec(asset.balance?.amount.toString() ?? 0, asset.decimal_digits);
-    const balance = formatTokenBalance(value);
-    const exactBalance = value.isZero() ? "0" : value.toString(asset.decimal_digits).replace(/\.?0+$/, "");
-    const price = new Dec(getPriceForCurrency(currency));
-    const stable = price.mul(value);
-
-    data.push({
-      name: asset.name,
-      value: denom,
-      label: asset.shortName,
-      shortName: asset.shortName,
-      icon: asset.icon,
-      decimal_digits: asset.decimal_digits,
-      balance: {
-        value: exactBalance,
-        customLabel: `${balance} ${asset.shortName}`,
-        ticker: asset.shortName,
-        denom: asset.balance.denom,
-        amount: asset.balance?.amount
-      },
-      ibcData: (asset as ExternalCurrency).ibcData,
-      from: (asset as AssetBalance).from,
-      native: asset.native,
-      symbol: asset.symbol,
-      ticker: asset.ticker,
-      stable,
-      price: formatDecAsUsd(stable)
-    });
-  }
-
-  return data.sort((a, b) => {
-    return Number(b.stable.sub(a.stable).toString(8));
-  });
-});
-
-const currency = computed(() => {
-  return assets.value[selectedCurrency.value];
-});
-
-let client: Wallet;
-let timeOut!: NodeJS.Timeout;
-let route: RouteResponse | null;
-
-const walletStore = useWalletStore();
-const balancesStore = useBalancesStore();
-const configStore = useConfigStore();
-const historyStore = useHistoryStore();
-const networks = ref<Network[]>(NETWORK_DATA.list);
-
-const selectedNetwork = ref(0);
-const networkCurrencies = ref<ExternalCurrency[] | AssetBalance[]>(balancesStore.filteredBalances);
-const selectedCurrency = ref(0);
-const amount = ref("");
-const amountErrorMsg = ref("");
-const errorInsufficientBalance = computed(() => amountErrorMsg.value === i18n.t("message.invalid-balance-big"));
-const txHashes = ref<{ hash: string; status: SwapStatus; url: string | null }[]>([]);
-
-const step = ref(CONFIRM_STEP.CONFIRM);
-const fee = ref<Coin>();
-const isLoading = ref(false);
-const disablePicker = ref(false);
-const isDisabled = ref(false);
-const receiverAddress = ref("");
-const tempRoute = ref<RouteResponse | null>(null);
-let chainsData: Chain[] = [];
-
-let skipRouteConfig: SkipRouteConfigType | null;
-const id = Date.now();
-const wallet = ref(walletStore.wallet?.address);
-const onClose = inject("close", () => {});
-
-const network = computed(() => {
-  return networks.value[selectedNetwork.value];
-});
-
-const networkCurrenciesRef = computed(() => {
-  if (network.value.native) {
-    return balancesStore.filteredBalances;
-  }
-  return networkCurrencies.value;
-});
-
-const walletRef = computed(() => {
-  if (network.value.native) {
-    return "";
-  }
-  return wallet.value;
-});
-
-// Repopulate when the store is ready AND whenever the wallet-driven network
-// filter changes. protocolFilter starts "" and is set to the owned network
-// (e.g. "OSMOSIS") only after the async wallet reconnect completes — later than
-// `initialized` flips. Reacting to it also lands the form on the destination
-// network so the Recipient resolves to the derived destination address
-// (e.g. osmo1…) instead of falling back to the native Nolus account.
-//
-// Reacts to: configStore.initialized + configStore.protocolFilter.
-// Idempotency: onInit() recomputes form/network state from the current store
-// snapshot, so the immediate run and every re-fire are safe.
-watch(
-  [() => configStore.initialized, () => configStore.protocolFilter],
-  () => {
-    if (configStore.initialized) {
-      onInit();
-    }
-  },
-  {
-    immediate: true
-  }
-);
-
-async function onInit() {
-  try {
-    const [config, chns] = await Promise.all([getSkipRouteConfig(), SkipRouter.getChains()]);
-    skipRouteConfig = config;
-    chainsData = chns;
-    const n = NETWORK_DATA.list.filter((item) => {
-      if (config.transfers[item.key]) {
-        return true;
-      }
-      return false;
-    }) as Network[];
-    networks.value = [...n].filter((item) => {
-      return !IGNORED_NETWORKS.includes(item.key);
-    });
-    const index = networks.value.findIndex((item: Network) => item.key === configStore.protocolFilter);
-    if (index < 0) {
-      selectedNetwork.value = 0;
-    } else {
-      selectedNetwork.value = index;
-    }
-    await onUpdateNetwork(network.value);
-  } catch (error) {
-    Logger.error(error);
-  }
-}
-
-onUnmounted(() => {
-  clearTimeout(timeOut);
-  if (client && step.value !== CONFIRM_STEP.PENDING) {
-    destroyClient();
-  }
-});
-
-const calculatedBalance = computed(() => {
-  const asset = assets.value[selectedCurrency.value];
-  if (!asset) {
-    return formatUsd(0);
-  }
-  const denom = asset.ibcData ?? asset.from;
-  const currency = tryGetCurrencyByDenom(denom);
-  if (!currency) return formatUsd(0);
-
-  const price = new Dec(getPriceForCurrency(currency));
-  const v = amount?.value?.length ? amount?.value : "0";
-  const stable = price.mul(new Dec(v));
-  return formatDecAsUsd(stable);
-});
-
-function destroyClient() {
-  try {
-    client.destroy();
-  } catch {
-    // intentionally empty - destroy errors are non-critical
-  }
-}
-
-function setHistory() {
-  if (!tempRoute.value) return;
-  const chains = getChainIds(tempRoute.value);
-
-  const data = {
-    id,
-    chains,
-    skipRoute: route,
-    currency: currency.value.from,
-    fromAddress: walletStore.wallet?.address,
-    receiverAddress: wallet.value,
-    type: HISTORY_ACTIONS.SEND
-  };
-  historyStore.addPendingTransfer(data, i18n);
-}
-
-// Reacts to: selected currency, amount, and wallet address.
-// Idempotency: validateAmount() runs first (resets amountErrorMsg each call) and
-// gates a debounced route fetch; the clearTimeout collapses rapid re-fires into a
-// single getRoute() call.
-watch(
-  () => [selectedCurrency.value, amount.value, wallet.value],
-  () => {
-    if (amount.value.length > 0) {
-      if (validateAmount() && wallet.value?.length > 0) {
-        clearTimeout(timeOut);
-        tempRoute.value = null;
-        timeOut = setTimeout(async () => {
-          try {
-            tempRoute.value = await getRoute();
-          } catch (e: unknown) {
-            amountErrorMsg.value = i18n.t(classifyError(e));
-          }
-        });
-      }
-    }
-  }
-);
-
-// Reacts to: receiverAddress (user input).
-// Idempotency: recomputes amountErrorMsg via validateAddress on each change; pure
-// validation, safe to re-fire.
-watch(
-  () => receiverAddress.value,
-  () => {
-    amountErrorMsg.value = validateAddress(receiverAddress.value);
-  }
-);
-
-async function onUpdateNetwork(event: Network) {
-  tempRoute.value = null;
-  selectedNetwork.value = networks.value.findIndex((item) => item === event);
-  if (!event.native) {
-    await setCosmosNetwork();
-  } else {
-    setNativeNetwork();
-  }
-}
-
-function handleAmountChange(event: string) {
-  amount.value = event;
-}
-
-async function onSubmitCosmos() {
-  try {
-    isDisabled.value = true;
-    amountErrorMsg.value = "";
-
-    const isValid = validateAmount();
-
-    if (isValid) {
-      route = await getRoute();
-      await onSwap();
-    }
-  } catch (e: unknown) {
-    amountErrorMsg.value = i18n.t(classifyError(e));
-  } finally {
-    isDisabled.value = false;
-  }
-}
-
-function setNativeNetwork() {
-  amountErrorMsg.value = "";
-  route = null;
-
-  networkCurrencies.value = balancesStore.filteredBalances;
-  selectedCurrency.value = 0;
-}
-
-async function setCosmosNetwork() {
-  networkCurrencies.value = [];
-  amount.value = "";
-  amountErrorMsg.value = "";
-  destroyClient();
-
-  disablePicker.value = true;
-  try {
-    const ntwrk = NETWORK_DATA;
-    const currencies = [];
-    const data = skipRouteConfig?.transfers?.[network.value.key]?.currencies;
-    for (const c of data ?? []) {
-      if (c.visible && configStore.protocolFilter !== c.visible) continue;
-
-      const currency = tryGetCurrencyByDenom(c.from);
-      if (!currency) continue; // Skip deprecated/removed denoms
-
-      const balance = balancesStore.getBalanceInfo(c.from);
-      currency.balance = coin(balance?.amount?.toString() ?? 0, c.to);
-      currencies.push(currency);
-    }
-
-    const mappedCurrencies = currencies.map((item) => {
-      return {
-        balance: item.balance,
-        shortName: item.shortName,
-        ticker: item.ticker,
-        name: item.shortName,
-        icon: item.icon,
-        decimal_digits: item.decimal_digits,
-        symbol: item.symbol,
-        native: item.native,
-        from: item.ibcData
-      };
-    });
-    const networkData = ntwrk?.supportedNetworks[network.value.key];
-    client = await WalletAccess.getWallet(network.value.key);
-    const baseWallet = (await externalWallet(client, networkData)) as BaseWallet;
-    wallet.value = baseWallet?.address as string;
-
-    networkCurrencies.value = mappedCurrencies;
-    selectedCurrency.value = 0;
-  } finally {
-    disablePicker.value = false;
-  }
-}
-
-function validateAmount() {
-  amountErrorMsg.value = "";
-
-  if (!WalletAccess.isAuth()) {
-    return false;
-  }
-
-  if (amount.value.at(0) === ".") {
-    amountErrorMsg.value = i18n.t("message.invalid-amount");
-    return false;
-  }
-
-  if (!amount.value) {
-    amountErrorMsg.value = i18n.t("message.invalid-amount");
-    return false;
-  }
-  const asset = assets.value[selectedCurrency.value];
-
-  const decimals = asset?.decimal_digits;
-  if (decimals) {
-    try {
-      const balance = asset.balance;
-      const walletBalance = Decimal.fromAtomics(balance.amount.toString(), decimals);
-      const transferAmount = Decimal.fromUserInput(amount.value, decimals);
-      const isGreaterThanWalletBalance = transferAmount.isGreaterThan(walletBalance);
-      const isLowerThanOrEqualsToZero = transferAmount.isLessThanOrEqual(Decimal.fromUserInput("0", decimals));
-
-      if (isLowerThanOrEqualsToZero) {
-        amountErrorMsg.value = i18n.t("message.invalid-balance-low");
-        return false;
-      }
-
-      if (isGreaterThanWalletBalance) {
-        amountErrorMsg.value = i18n.t("message.invalid-balance-big");
-        return false;
-      }
-    } catch (e) {
-      Logger.error(e);
-    }
-  } else {
-    amountErrorMsg.value = i18n.t("message.unexpected-error");
-    return false;
-  }
-
-  return true;
-}
-
-async function onSendClick() {
-  if (network.value.native) {
-    return onSubmitNative();
-  }
-  onSubmitCosmos();
-}
-
-function onSubmitNative() {
-  const isValid = validateAmount() && validateAddress(receiverAddress.value).length === 0;
-
-  if (isValid) {
-    onSwap();
-  }
-}
-
-async function onSwap() {
-  try {
-    if (network.value.native) {
-      await onSwapNative();
-    } else {
-      await onSwapCosmos();
-    }
-  } catch {
-    step.value = CONFIRM_STEP.ERROR;
-  }
-}
-
-async function onSwapNative() {
-  try {
-    isDisabled.value = true;
-    await walletOperation(transferAmount);
-  } catch {
-    step.value = CONFIRM_STEP.ERROR;
-  } finally {
-    isDisabled.value = false;
-  }
-}
-
-async function transferAmount() {
-  try {
-    isLoading.value = true;
-    step.value = CONFIRM_STEP.PENDING;
-    const asset = assets.value[selectedCurrency.value];
-    const { success, txHash, txBytes, usedFee } = await transferCurrency(
-      asset.balance.denom,
-      amount.value,
-      receiverAddress.value,
-      ""
-    );
-
-    if (success) {
-      const element = {
-        hash: txHash,
-        status: SwapStatus.pending,
-        url: null as string
-      };
-
-      const index = txHashes.value.length;
-      txHashes.value.push(element);
-
-      if (usedFee?.amount?.[0]) {
-        fee.value = usedFee.amount[0];
-      }
-
-      try {
-        const tx = await walletStore.wallet?.broadcastTx(txBytes as Uint8Array);
-        const isSuccessful = tx?.code === 0;
-        step.value = isSuccessful ? CONFIRM_STEP.SUCCESS : CONFIRM_STEP.ERROR;
-        txHashes.value[index].status = SwapStatus.success;
-
-        await balancesStore.fetchBalances();
-      } catch (error: unknown) {
-        if (walletStore.history[id]) {
-          walletStore.history[id].historyData.errorMsg = amountErrorMsg.value;
-          walletStore.history[id].historyData.status = CONFIRM_STEP.ERROR;
-        }
-        switch ((error as { code?: number }).code) {
-          case ErrorCodes.GasError: {
-            step.value = CONFIRM_STEP.GasError;
-            break;
-          }
-          default: {
-            step.value = CONFIRM_STEP.ERROR;
-            break;
-          }
-        }
-      } finally {
-        destroyClient();
-      }
-    } else {
-      step.value = CONFIRM_STEP.ERROR;
-
-      if (walletStore.history[id]) {
-        walletStore.history[id].historyData.errorMsg = amountErrorMsg.value;
-        walletStore.history[id].historyData.status = CONFIRM_STEP.ERROR;
-      }
-    }
-
-    historyStore.loadActivities();
-    onClose();
-  } catch (e: unknown) {
-    amountErrorMsg.value = i18n.t(classifyError(e));
-  } finally {
-    isLoading.value = false;
-  }
-}
-
-async function onSwapCosmos() {
-  if (!route || !WalletAccess.isAuth() || amountErrorMsg.value.length > 0) {
-    return false;
-  }
-
-  try {
-    step.value = CONFIRM_STEP.PENDING;
-    isDisabled.value = true;
-
-    await walletOperation(async () => {
-      try {
-        isLoading.value = true;
-        const wallets = await getWallets();
-        const addresses: Record<string, string> = {};
-
-        for (const key in wallets) {
-          const walletAddress = wallets[key].address;
-          if (!walletAddress) {
-            throw new Error(`Wallet address not available for ${key}`);
-          }
-          addresses[key] = walletAddress;
-        }
-
-        setHistory();
-        await submit(wallets);
-        await balancesStore.fetchBalances();
-
-        historyStore.loadActivities();
-
-        step.value = CONFIRM_STEP.SUCCESS;
-        walletStore.history[id].historyData.route.activeStep = walletStore.history[id].historyData.route.steps.length;
-        walletStore.history[id].historyData.routeDetails.activeStep =
-          walletStore.history[id].historyData.routeDetails.steps.length;
-        walletStore.history[id].historyData.status = CONFIRM_STEP.SUCCESS;
-
-        onClose();
-      } catch (error) {
-        step.value = CONFIRM_STEP.ERROR;
-        amountErrorMsg.value = i18n.t(classifyError(error));
-        Logger.error(error);
-
-        if (walletStore.history[id]) {
-          walletStore.history[id].historyData.errorMsg = amountErrorMsg.value;
-          walletStore.history[id].historyData.route.steps[
-            walletStore.history[id].historyData.route.activeStep
-          ].approval = true;
-
-          walletStore.history[id].historyData.routeDetails.steps[
-            walletStore.history[id].historyData.routeDetails.activeStep
-          ].approval = true;
-
-          walletStore.history[id].historyData.status = CONFIRM_STEP.ERROR;
-        }
-      } finally {
-        isLoading.value = false;
-      }
-    });
-  } catch (e) {
-    Logger.error(e);
-  } finally {
-    isDisabled.value = false;
-    destroyClient();
-  }
-}
-
-async function submit(wallets: { [key: string]: BaseWallet }) {
-  if (!route) {
-    throw new Error("Route not available");
-  }
-  await SkipRouter.submitRoute(route, wallets, async (tx: SkipTxResult, wallet: BaseWallet, chainId: string) => {
-    walletStore.history[id].historyData.route.activeStep++;
-    walletStore.history[id].historyData.routeDetails.activeStep++;
-
-    const element = {
-      hash: tx.txHash,
-      status: SwapStatus.pending,
-      url: wallet.explorer
-    };
-
-    txHashes.value.push(element);
-
-    if (walletStore.history[id]) {
-      walletStore.history[id].historyData.txHashes = txHashes.value;
-    }
-
-    await wallet.broadcastTx(tx.txBytes);
-    await SkipRouter.track(chainId, tx.txHash);
-    await SkipRouter.fetchStatus(tx.txHash, chainId);
-
-    element.status = SwapStatus.success;
-  });
-}
-
-async function getWallets(): Promise<{ [key: string]: BaseWallet }> {
-  if (!route) {
-    throw new Error("Route not available");
-  }
-  const native = walletStore.wallet.signer.chainId as string;
-  const addrs = {
-    [native]: walletStore.wallet
-  };
-
-  const chainToParse: { [key: string]: NetworkInfo } = getChains(route);
-  const promises = [];
-  for (const chain in chainToParse) {
-    const fn = async function () {
-      if (chain !== NATIVE_NETWORK.key) {
-        const client = await WalletAccess.getWallet(chain);
-        const network = NETWORK_DATA;
-        const networkData = network?.supportedNetworks[chain];
-        const baseWallet = (await externalWallet(client, networkData)) as BaseWallet;
-        const chainId = await baseWallet.getChainId();
-        addrs[chainId] = baseWallet;
-      }
-    };
-    promises.push(fn());
-  }
-
-  await Promise.all(promises);
-  return addrs;
-}
-
-async function getRoute() {
-  const chainId = await client.getChainId();
-  const asset = assets.value[selectedCurrency.value];
-
-  const transferAmount = Decimal.fromUserInput(amount.value, asset.decimal_digits as number);
-
-  const route = await SkipRouter.getRoute(
-    asset.from,
-    asset.balance.denom,
-    transferAmount.atomics,
-    false,
-    undefined,
-    chainId
-  );
-
-  return route;
-}
-
-function getChains(route: RouteResponse) {
-  const chainToParse: { [key: string]: NetworkInfo } = {};
-  const native = walletStore.wallet.signer.chain_id as string;
-  const chains = chainsData.filter((item) => {
-    if (item.chain_id === native) {
-      return false;
-    }
-    return route.chain_ids?.includes(item.chain_id);
-  });
-
-  const supportedNetworks = configStore.supportedNetworksData;
-  for (const chain of chains) {
-    for (const key in supportedNetworks) {
-      const networkData = supportedNetworks[key];
-      if (networkData?.value === chain.chain_name.toLowerCase()) {
-        chainToParse[key] = networkData;
-      }
-    }
-  }
-
-  return chainToParse;
-}
-
-function getChainIds(route: RouteResponse) {
-  const chainToParse: { [key: string]: NetworkInfo } = {};
-  const chains = chainsData.filter((item) => {
-    return route.chain_ids?.includes?.(item.chain_id);
-  });
-
-  const supportedNetworks = configStore.supportedNetworksData;
-  for (const chain of chains) {
-    for (const key in supportedNetworks) {
-      const networkData = supportedNetworks[key];
-      if (networkData?.value === chain.chain_name.toLowerCase()) {
-        chainToParse[chain.chain_id] = networkData;
-      }
-    }
-  }
-
-  return chainToParse;
-}
+import { useSendForm } from "./useSendForm";
+
+const {
+  assets,
+  currency,
+  amount,
+  amountErrorMsg,
+  errorInsufficientBalance,
+  calculatedBalance,
+  disablePicker,
+  isDisabled,
+  isLoading,
+  network,
+  walletRef,
+  handleAmountChange,
+  onSelectCurrency,
+  onRecipientInput,
+  onSendClick
+} = useSendForm();
 </script>
